requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# ── Ingestão ─────────────────────────────────────────────────────────────────
requests==2.32.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
openpyxl==3.1.2          # leitura de .xlsx

# ── Transformação ─────────────────────────────────────────────────────────────
tabulate==0.9.0           # df.to_markdown()

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.5.0
tensorflow==2.16.1

# ── IA & Agentes ──────────────────────────────────────────────────────────────
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-core==0.2.3

# ── Orquestração ──────────────────────────────────────────────────────────────
apache-airflow==2.9.1

# ── Notebooks ────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# ── Qualidade de dados ────────────────────────────────────────────────────────
great-expectations==0.18.15

# ── Testes ───────────────────────────────────────────────────────────────────
pytest==8.2.1
pytest-cov==5.0.0
